:recycle: Refactor Address component to typescript
