--- src/js/shop/components/Checkout/Context.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const CheckoutContext = React.createContext({
-    validationErrors: null,
-    customer: null,
-    deliveryAddress: null,
-    billingAddress: null,
-});
-
-CheckoutContext.displayName = "CheckoutContext";
-
-export { CheckoutContext };
